Refactor so that Dates is stateful again, and bookings are loaded there
